+++ src/modules/explorer/v2/pages/Proposals/index.tsx
@@ -1,0 +1,96 @@
+import React, { useCallback } from "react";
+import { Grid, styled, Typography, Button } from "@material-ui/core";
+
+import { useFlush } from "services/contracts/baseDAO/hooks/useFlush";
+import { useDAO } from "services/indexer/dao/hooks/useDAO";
+import { useProposals } from "services/indexer/dao/hooks/useProposals";
+import { useDAOID } from "../DAO/router";
+
+import { UserBalancesBox } from "../../components/UserBalances";
+import { ContentContainer } from "../../components/ContentContainer";
+import { ProposalsList } from "../../components/ProposalsList";
+import { DAOStatsRow } from "../../components/DAOStatsRow";
+import { ProposalStatus } from "services/indexer/dao/mappers/proposal/types";
+
+const HeroContainer = styled(ContentContainer)({
+  padding: "38px 45px",
+});
+
+const TitleText = styled(Typography)({
+  fontSize: 30,
+  fontWeight: 500,
+});
+
+export const Proposals: React.FC = () => {
+  const daoId = useDAOID();
+  const { data, cycleInfo } = useDAO(daoId);
+  const { mutate } = useFlush();
+
+  console.log(cycleInfo)
+
+  const { data: proposals } = useProposals(daoId);
+  const { data: activeProposals } = useProposals(daoId, ProposalStatus.ACTIVE);
+
+  const onFlush = useCallback(async () => {
+    if (proposals && proposals.length && data) {
+      mutate({
+        dao: data,
+        numOfProposalsToFlush: proposals.length + 1,
+      });
+      return;
+    }
+  }, [data, mutate, proposals]);
+
+  return (
+    <Grid container direction="column" style={{ gap: 42 }}>
+      <HeroContainer item>
+        <Grid container justifyContent="space-between">
+          <Grid item>
+            <Grid container style={{ gap: 20 }} alignItems="center">
+              <Grid item>
+                <TitleText color="textPrimary">Proposals</TitleText>
+              </Grid>
+              <Grid item>
+                <Button
+                  variant="contained"
+                  color="secondary"
+                  size="small"
+                  onClick={onFlush}
+                >
+                  Execute
+                </Button>
+              </Grid>
+            </Grid>
+          </Grid>
+          <Grid item>
+            <Button
+              variant="contained"
+              color="secondary"
+              onClick={onFlush}
+            >
+              New Proposal
+            </Button>
+          </Grid>
+        </Grid>
+      </HeroContainer>
+      <DAOStatsRow />
+      <UserBalancesBox daoId={daoId} />
+
+      {data && cycleInfo && activeProposals && (
+        <ProposalsList
+          title={"Active Proposals"}
+          currentLevel={cycleInfo.currentLevel}
+          proposals={activeProposals}
+        />
+      )}
+
+      {data && cycleInfo && proposals && (
+        <ProposalsList
+          title={"All Proposals"}
+          currentLevel={cycleInfo.currentLevel}
+          proposals={proposals}
+        />
+      )}
+    </Grid>
+  );
+};
